Make only right side panel scroll. More work on integrating components with app.
